+++ plugins/org.preesm.model.slam/src/org/preesm/model/slam/route/SlamRoutingTable.java
@@ -1,10 +1,10 @@
 /**
- * Copyright or © or Copr. IETR/INSA - Rennes (2008 - 2019) :
+ * Copyright or © or Copr. IETR/INSA - Rennes (2009 - 2019) :
  *
  * Antoine Morvan [antoine.morvan@insa-rennes.fr] (2017 - 2019)
- * Clément Guy [clement.guy@insa-rennes.fr] (2014 - 2015)
- * Matthieu Wipliez [matthieu.wipliez@insa-rennes.fr] (2008)
- * Maxime Pelcat [maxime.pelcat@insa-rennes.fr] (2008 - 2014)
+ * Clément Guy [clement.guy@insa-rennes.fr] (2014)
+ * Jonathan Piat [jpiat@laas.fr] (2011)
+ * Maxime Pelcat [maxime.pelcat@insa-rennes.fr] (2009 - 2016)
  *
  * This software is a computer program whose purpose is to help prototyping
  * parallel applications using dataflow formalism.
@@ -42,8 +42,11 @@ import java.util.LinkedHashMap;
 import java.util.LinkedHashSet;
 import java.util.List;
 import java.util.Map;
+import java.util.Map.Entry;
 import java.util.Set;
+import java.util.concurrent.ConcurrentSkipListSet;
 import java.util.logging.Level;
+import org.apache.commons.lang3.tuple.MutablePair;
 import org.preesm.commons.exceptions.PreesmRuntimeException;
 import org.preesm.commons.logger.PreesmLogger;
 import org.preesm.model.slam.ComponentInstance;
@@ -51,87 +54,156 @@ import org.preesm.model.slam.Design;
 import org.preesm.model.slam.Link;
 import org.preesm.model.slam.Operator;
 import org.preesm.model.slam.SlamRoute;
-import org.preesm.model.slam.SlamRouteStep;
 import org.preesm.model.slam.impl.ComNodeImpl;
 import org.preesm.model.slam.utils.SlamUserFactory;
 
 /**
- * This class can evaluate a given transfer and choose the best route between two operators.
+ * Table representing the different routes available to go from one operator to another.
  *
  * @author mpelcat
  */
-public class RouteCalculator {
-
-  /** The instances. */
-  private static Map<Design, RouteCalculator> instances = new LinkedHashMap<>();
-
-  /** The archi. */
-  private final Design archi;
-
-  /** The table. */
-  private RoutingTable table = null;
-
-  /** The scenario. */
+public class SlamRoutingTable {
 
   /**
-   * Gets the single instance of RouteCalculator.
-   *
+   * A couple of operators to which the routes are linked. Used a the key in Maps
    */
-  public static RouteCalculator getInstance(final Design archi, final long averageDataSize) {
-    if (RouteCalculator.instances.get(archi) == null) {
-      RouteCalculator.instances.put(archi, new RouteCalculator(archi, averageDataSize));
+  private class OperatorCouple extends MutablePair<ComponentInstance, ComponentInstance> {
+    private static final long serialVersionUID = -451571160460519876L;
+
+    OperatorCouple(final ComponentInstance op1, final ComponentInstance op2) {
+      super(op1, op2);
     }
-    return RouteCalculator.instances.get(archi);
+
+    @Override
+    public String toString() {
+      return "(" + getLeft() + "," + getRight() + ")";
+    }
   }
 
   /**
-   * Recalculate.
-   *
+   * A list of routes ordered in inverse order of transfer cost.
    */
-  public static void recalculate(final Design archi, final long averageDataSize) {
-    RouteCalculator.instances.put(archi, new RouteCalculator(archi, averageDataSize));
+  private class RouteList extends ConcurrentSkipListSet<SlamRoute> {
+
+    private static final long serialVersionUID = -851695207011182681L;
+
+    RouteList() {
+      super((o1, o2) -> {
+        final double difference = RouteCostEvaluator.evaluateTransferCost(o1, 1)
+            - RouteCostEvaluator.evaluateTransferCost(o2, 1);
+        if (difference >= 0) {
+          return 1;
+        } else {
+          return -1;
+        }
+      });
+    }
+
+    @Override
+    public String toString() {
+      final StringBuilder sb = new StringBuilder("|");
+      for (final SlamRoute r : this) {
+        sb.append(r.toString() + "|");
+      }
+      return sb.toString();
+    }
   }
 
-  /**
-   * Delete routes.
-   *
-   * @param archi
-   *          the archi
-   */
-  public static void deleteRoutes(final Design archi) {
-    RouteCalculator.instances.remove(archi);
-  }
+  /** List of available routes. */
+  private final Map<OperatorCouple, RouteList> table;
+  private final Design                         archi;
 
   /**
-   * Constructor from a given architecture.
+   * Instantiates a new routing table.
    *
-   * @param archi
-   *          the archi
-   * @param scenario
-   *          the scenario
    */
-  private RouteCalculator(final Design archi, final long averageDataSize) {
-
+  public SlamRoutingTable(final Design archi) {
+    super();
     this.archi = archi;
-    this.table = new RoutingTable();
+    this.table = new LinkedHashMap<>();
 
     // Creating the route steps between directly connected operators
-    createRouteSteps(averageDataSize);
+    createRouteSteps();
     // Concatenation of route steps to generate optimal routes using
     // the Floyd Warshall algorithm
-    createRoutes(averageDataSize);
+    createRoutes();
+  }
+
+  /**
+   * Gets a route with a given index.
+   *
+   * @param op1
+   *          the op 1
+   * @param op2
+   *          the op 2
+   * @return the best route
+   */
+  SlamRoute getBestRoute(final ComponentInstance op1, final ComponentInstance op2) {
+    final OperatorCouple obj = new OperatorCouple(op1, op2);
+    if (this.table.containsKey(obj)) {
+      return this.table.get(obj).first();
+    }
+    return null;
+  }
+
+  /**
+   * Removes all the routes corresponding to the operator couple.
+   *
+   * @param op1
+   *          the op 1
+   * @param op2
+   *          the op 2
+   */
+  void removeRoutes(final ComponentInstance op1, final ComponentInstance op2) {
+    final OperatorCouple route = new OperatorCouple(op1, op2);
+    if (this.table.containsKey(route)) {
+      this.table.get(route).clear();
+    }
+  }
+
+  /**
+   * Adds a new route.
+   *
+   * @param op1
+   *          the op 1
+   * @param op2
+   *          the op 2
+   * @param route
+   *          the route
+   */
+  void addRoute(final ComponentInstance op1, final ComponentInstance op2, final SlamRoute route) {
+    final OperatorCouple opCouple = new OperatorCouple(op1, op2);
+    if (!this.table.containsKey(opCouple)) {
+      this.table.put(opCouple, new RouteList());
+    }
+    this.table.get(opCouple).add(route);
+  }
+
+  /**
+   * Displays the table.
+   *
+   * @return the string
+   */
+  @Override
+  public String toString() {
+    final StringBuilder sb = new StringBuilder();
+    for (final Entry<OperatorCouple, RouteList> e : this.table.entrySet()) {
+      final OperatorCouple couple = e.getKey();
+      sb.append(couple + " -> " + e.getValue() + "\n");
+    }
+    return sb.toString();
   }
 
   /**
    * Creating recursively the route steps from the architecture.
    */
-  private void createRouteSteps(final long averageDataSize) {
+  private void createRouteSteps() {
     PreesmLogger.getLogger().log(Level.INFO, "creating route steps.");
 
     for (final ComponentInstance c : this.archi.getOperatorComponentInstances()) {
       final ComponentInstance o = c;
 
-      createRouteSteps(o, averageDataSize);
+      createRouteSteps(o);
     }
   }
 
@@ -141,7 +213,7 @@ public class RouteCalculator {
    * @param source
    *          the source
    */
-  private void createRouteSteps(final ComponentInstance source, final long averageDataSize) {
+  private void createRouteSteps(final ComponentInstance source) {
 
     // Iterating on outgoing and undirected edges
     final Set<Link> outgoingAndUndirected = new LinkedHashSet<>();
@@ -156,7 +228,7 @@ public class RouteCalculator {
 
         final List<ComponentInstance> alreadyVisitedNodes = new ArrayList<>();
         alreadyVisitedNodes.add(node);
-        exploreRoute(source, node, alreadyVisitedNodes, averageDataSize);
+        exploreRoute(source, node, alreadyVisitedNodes);
       }
     }
   }
@@ -172,7 +244,7 @@ public class RouteCalculator {
    *          the already visited nodes
    */
   private void exploreRoute(final ComponentInstance source, final ComponentInstance node,
-      final List<ComponentInstance> alreadyVisitedNodes, final long avgSize) {
+      final List<ComponentInstance> alreadyVisitedNodes) {
 
     // Iterating on outgoing and undirected edges
     final Set<Link> outgoingAndUndirected = new LinkedHashSet<>();
@@ -186,14 +258,13 @@ public class RouteCalculator {
         if (!alreadyVisitedNodes.contains(newNode)) {
           final List<ComponentInstance> newAlreadyVisitedNodes = new ArrayList<>(alreadyVisitedNodes);
           newAlreadyVisitedNodes.add(newNode);
-          exploreRoute(source, newNode, newAlreadyVisitedNodes, avgSize);
+          exploreRoute(source, newNode, newAlreadyVisitedNodes);
         }
       } else if ((otherEnd.getComponent() instanceof Operator)
           && !otherEnd.getInstanceName().equals(source.getInstanceName())) {
-        final ComponentInstance target = otherEnd;
-        final SlamRouteStep step = SlamUserFactory.eINSTANCE.createRouteStep(this.archi, source, alreadyVisitedNodes,
-            target);
-        this.table.addRoute(source, target, SlamUserFactory.eINSTANCE.createSlamRoute(step), avgSize);
+        final SlamRoute newRoute = SlamUserFactory.eINSTANCE.createSlamRoute(this.archi, source, alreadyVisitedNodes,
+            otherEnd);
+        this.addRoute(source, otherEnd, newRoute);
       }
     }
   }
@@ -201,10 +272,10 @@ public class RouteCalculator {
   /**
    * Building recursively the routes between the cores.
    */
-  private void createRoutes(final long avgSize) {
+  private void createRoutes() {
     PreesmLogger.getLogger().log(Level.INFO, "Initializing routing table.");
 
-    floydWarshall(this.table, this.archi.getOperatorComponentInstances(), avgSize);
+    floydWarshall(this, this.archi.getOperatorComponentInstances());
   }
 
   /**
@@ -215,8 +286,7 @@ public class RouteCalculator {
    * @param operators
    *          the operators
    */
-  private void floydWarshall(final RoutingTable table, final List<ComponentInstance> operators,
-      final long averageDataSize) {
+  private void floydWarshall(final SlamRoutingTable table, final List<ComponentInstance> operators) {
 
     for (final ComponentInstance k : operators) {
 
@@ -235,13 +305,13 @@ public class RouteCalculator {
                 // routes become available
                 final SlamRoute bestRoute = table.getBestRoute(src, tgt);
                 if (bestRoute == null) {
-                  table.addRoute(src, tgt, compoundRoute, averageDataSize);
+                  table.addRoute(src, tgt, compoundRoute);
                 } else {
-                  final long bestRouteCost = RouteCostEvaluator.evaluateTransferCost(bestRoute, averageDataSize);
-                  final long newRouteCost = RouteCostEvaluator.evaluateTransferCost(compoundRoute, averageDataSize);
+                  final double bestRouteCost = RouteCostEvaluator.evaluateTransferCost(bestRoute, 1);
+                  final double newRouteCost = RouteCostEvaluator.evaluateTransferCost(compoundRoute, 1);
                   if (bestRouteCost > newRouteCost) {
                     table.removeRoutes(src, tgt);
-                    table.addRoute(src, tgt, compoundRoute, averageDataSize);
+                    table.addRoute(src, tgt, compoundRoute);
                   }
                 }
               }
@@ -262,7 +332,7 @@ public class RouteCalculator {
    * @return the route
    */
   public SlamRoute getRoute(final ComponentInstance op1, final ComponentInstance op2) {
-    final SlamRoute r = this.table.getBestRoute(op1, op2);
+    final SlamRoute r = this.getBestRoute(op1, op2);
 
     if (r == null) {
       final String msg = "Did not find a route between " + op1 + " and " + op2 + ".";