+++ plugins/org.preesm.model.slam/src/org/preesm/model/slam/utils/SlamUserFactory.java
@@ -47,7 +47,6 @@ import org.preesm.model.slam.SlamRoute;
 import org.preesm.model.slam.SlamRouteStep;
 import org.preesm.model.slam.VLNV;
 import org.preesm.model.slam.impl.SlamFactoryImpl;
-import org.preesm.model.slam.route.RouteStepFactory;
 
 /**
  *
@@ -95,19 +94,25 @@ public class SlamUserFactory extends SlamFactoryImpl {
     return res;
   }
 
+  public SlamRoute createSlamRoute(final Design archi, final ComponentInstance source,
+      final List<ComponentInstance> alreadyVisitedNodes, final ComponentInstance target) {
+    return createSlamRoute(createSlamRouteStep(archi, source, alreadyVisitedNodes, target));
+  }
+
   /**
    *
    */
-  public final SlamRouteStep createRouteStep(final Design archi, final ComponentInstance source,
+  public final SlamRouteStep createSlamRouteStep(final Design archi, final ComponentInstance source,
       final List<ComponentInstance> nodes, final ComponentInstance target) {
     SlamRouteStep step = null;
 
-    final ComponentInstance dma = RouteStepFactory.getDma(archi, nodes, source);
-    final ComponentInstance mem = RouteStepFactory.getRam(archi, nodes, source);
+    final ComponentInstance dma = SlamCommunicationFinder.getDma(archi, nodes, source);
+    final ComponentInstance mem = SlamCommunicationFinder.getRam(archi, nodes, source);
     if (dma != null) {
       step = createSlamDMARouteStep(source, nodes, target, dma);
     } else if (mem != null) {
-      step = createSlamMemoryRouteStep(source, nodes, target, mem, RouteStepFactory.getRamNodeIndex(archi, nodes));
+      step = createSlamMemoryRouteStep(source, nodes, target, mem,
+          SlamCommunicationFinder.getRamNodeIndex(archi, nodes));
     } else {
       step = createSlamMessageRouteStep(source, nodes, target);
     }
@@ -156,4 +161,5 @@ public class SlamUserFactory extends SlamFactoryImpl {
     res.getNodes().addAll(nodes);
     return res;
   }
+
 }